### START OF THE PIXEL PICKAXE PROJECT ###
This project is about having a massive multiplayer pixel art creation site where people would be
able to draw on a wall.
One of the concepts would be a limited set of colors: the web safe colors.
These will be 12 bit colors instead of 24 bit "real" colors. This way, it will take less network
ressources, and I will use the last 4 bits for the state of the pixel (selected,etc...)